Split code into 4 files.
