+++ source/pip/qsharp/_device/_atom/__init__.py
@@ -4,7 +4,7 @@
 from .._device import Device, Zone, ZoneType
 from ..._simulation import NoiseConfig, run_qir_clifford, run_qir_cpu, run_qir_gpu
 from ..._native import try_create_gpu_adapter
-from ..._qsharp import QirInputData
+from ..._types import QirInputData
 from ... import telemetry_events
 
 from typing import List, Literal, Optional
