+++ source/pip/qsharp/openqasm/_circuit.py
@@ -2,17 +2,16 @@
 # Licensed under the MIT License.
 
 from time import monotonic
-from typing import Any, Callable, Dict, Optional, Union
-from .._fs import read_file, list_directory, resolve
+from typing import Any, Callable, Optional, Union
+
+from .. import telemetry_events
+from .._fs import list_directory, read_file, resolve
 from .._http import fetch_github
-from .._native import circuit_qasm_program  # type: ignore
+from .._native import Circuit, CircuitConfig, circuit_qasm_program  # type: ignore 
 from .._qsharp import (
     _get_session,
     ipython_helper,
-    Circuit,
-    CircuitConfig,
 )
-from .. import telemetry_events
 
 
 def circuit(
